pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "apps-generator"
version = "0.1.0"
description = "Generate full-stack projects from configurable templates"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "pydantic>=2.5",
    "httpx>=0.27",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
appgen = "apps_generator.cli.main:run"

[tool.hatch.build.targets.wheel]
packages = ["src/apps_generator"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
